move tests into correct places, update initial seed list testing WIP
